src/demo/python/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
cdk-lambda-subminute
